requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.4
django-braces==1.11.0
django-filter==1.0.4
djangorestframework==3.6.3
dry-rest-permissions==0.1.9
freeze==1.0.10
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
psycopg2==2.7.3.1
py==1.4.34
Pygments==2.2.0
pytest==3.2.3
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
ansible==2.4.1
